Code refactory for HashAggregateExecBaseTransformer
